+++ optd-datafusion-repr/src/cost/base_cost/limit.rs
@@ -1,0 +1,50 @@
+use optd_core::{
+    cascades::{CascadesOptimizer, RelNodeContext},
+    cost::Cost,
+};
+
+use crate::{
+    cost::stats::{Distribution, MostCommonValues},
+    plan_nodes::{ConstantExpr, OptRelNodeTyp},
+};
+
+use super::{OptCostModel, DEFAULT_UNK_SEL};
+
+impl<M: MostCommonValues, D: Distribution> OptCostModel<M, D> {
+    pub(super) fn get_limit_cost(
+        children: &[Cost],
+        context: Option<RelNodeContext>,
+        optimizer: Option<&CascadesOptimizer<OptRelNodeTyp>>,
+    ) -> Cost {
+        let (row_cnt, compute_cost, _) = Self::cost_tuple(&children[0]);
+        let row_cnt = if let (Some(context), Some(optimizer)) = (context, optimizer) {
+            let mut fetch_expr =
+                optimizer.get_all_group_bindings(context.children_group_ids[2], false);
+            assert!(
+                fetch_expr.len() == 1,
+                "fetch expression should be the only expr in the group"
+            );
+            let fetch_expr = fetch_expr.pop().unwrap();
+            assert!(
+                matches!(
+                    fetch_expr.typ,
+                    OptRelNodeTyp::Constant(ConstantType::UInt64)
+                ),
+                "fetch type can only be UInt64"
+            );
+            let fetch = ConstantExpr::from_rel_node(fetch_expr)
+                .unwrap()
+                .value()
+                .as_u64();
+            // u64::MAX represents None
+            if fetch == u64::MAX {
+                row_cnt
+            } else {
+                row_cnt.min(fetch as f64)
+            }
+        } else {
+            (row_cnt * DEFAULT_UNK_SEL).max(1.0)
+        };
+        Self::cost(row_cnt, compute_cost, 0.0)
+    }
+}
